+++ framework/Routing/ControllerInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Framework\Routing;
+
+interface ControllerInterface
+{
+    /** Execute the controller logic */
+    public function execute(): void;
+}