+++ framework/Routing/ResourceControllerInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Framework\Routing;
+
+interface ResourceControllerInterface
+{
+}